Trying some React JS and creating a navigation bar
I have more plans to create a completely responsive website. I plan to make the navigation bar to interact with the mouse courser when hovering over. I will be committing changes to the navigation I am also learning React Bootstrap to make the navigation modernized. So far still a newbie with React so far and I am loving it!